services/api-gateway/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.30.0,<1
sqlalchemy>=2.0.36,<3
psycopg[binary]>=3.2.0,<4
alembic>=1.14.0,<2
bcrypt>=4.2.0,<5
PyJWT>=2.9.0,<3
python-multipart>=0.0.9,<1
email-validator>=2.0.0,<3
